New Data Structure
Changed how data was stored to allow for saving each iteration. Adds complementary code to parse the new data, but not fully implemented particularly for plotting. Also changed the folder name to be more accurate. Needs bug testing to see if method of saving works. Previous commit attempted to deal with random repeated failures in data collection due to invalid indices for number of scans (it was 0 or less somehow???), now takes one fewer scan than the number available and only if scans available is greater than 1