Kyle's app changes
Added functionality to all buttons concerning stage optimization, All functioned correctly with regards to the variables they change but need to be tested with the setup. changed references to app.stagecontrol to app.ex.PIstage. Added functionality to add scan buttons and made several edits to Modify all fields function, which could be used to simplify code for optimization table.