Kyle's app changes
Delete step button now works for the scan table, and most columns can now be edited. Ran into an issue when trying to edit the "Type" column which caused me to rethink how the scan information should be stored. Currently each new scan added increased the dimensions of the vectors in the scan structure rather than adding a dimension to the structure itself. This works fine on its own, but I expect it to cause problems when running scans so should be fixed before moving further.